// types/blocks.ts
export interface FutureOfIcon {
  iconCaption: string;
  icon: StrapiMediaObject;
  url: StrapiMediaObject;
}
export interface FutureOfColumn {
  title: string;
  futureOfIcons: FutureOfIcon[];
}
export interface OurSolutionItem {
  title: string;
  description: string; // HTML
}
export interface SolutionDescriptionGroup {
  ourSolutionItems: OurSolutionItem[];
}
export interface OurSolutionsBlock {
  __typename: "ComponentBlocksOurSolutions";
  heading?: string | null;
  description?: string | null; // HTML
  futureOfColum?: FutureOfColumn | null;
  ourSolutionDescription?: SolutionDescriptionGroup[] | null;
}